requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.0.2
langchain-milvus>=0.1.0
langchain-pinecone>=0.1.1
langchain-community>=0.2.0
pymilvus>=2.3.1
pydantic==2.4.2
pydantic[email]
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1 
requests==2.31.0
numpy>=1.22.0
minio>=7.2.0
pypdf